+++ src/vcode.c
@@ -51,7 +51,7 @@ DECLARE_AND_DEFINE_ARRAY(vcode_type);
     || x == VCODE_OP_CLOSURE || x == VCODE_OP_PROTECTED_INIT)
 #define OP_HAS_IDENT(x)                                                 \
    (x == VCODE_OP_LINK_VAR || x == VCODE_OP_LINK_PACKAGE                \
-    || x == VCODE_OP_DEBUG_LOCUS)
+    || x == VCODE_OP_DEBUG_LOCUS || x == VCODE_OP_LINK_INSTANCE)
 #define OP_HAS_REAL(x)                                                  \
    (x == VCODE_OP_CONST_REAL)
 #define OP_HAS_VALUE(x)                                                 \
@@ -906,7 +906,7 @@ const char *vcode_op_string(vcode_op_t op)
       "debug locus", "length check", "range check", "array ref", "range length",
       "exponent check", "zero check", "map const", "resolve signal",
       "push scope", "pop scope", "set signal kind", "alias signal", "trap add",
-      "trap sub", "trap mul", "force", "release"
+      "trap sub", "trap mul", "force", "release", "link instance"
    };
    if ((unsigned)op >= ARRAY_LEN(strs))
       return "???";
@@ -2094,16 +2094,21 @@ void vcode_dump_with_mark(int mark_op, vcode_dump_fn_t callback, void *arg)
                col += printf(" right ");
                vcode_dump_reg(op->args.items[1]);
                col += printf(" dir ");
-               vcode_dump_reg(op->args.items[2]);
+               col += vcode_dump_reg(op->args.items[2]);
                vcode_dump_result_type(col, op);
             }
             break;
 
          case VCODE_OP_LINK_PACKAGE:
+         case VCODE_OP_LINK_INSTANCE:
             {
                col += vcode_dump_reg(op->result);
                col += color_printf(" := %s $magenta$%s$$",
                                    vcode_op_string(op->kind), istr(op->ident));
+               if (op->args.count > 0) {
+                  col += printf(" locus ");
+                  col += vcode_dump_reg(op->args.items[0]);
+               }
                vcode_dump_result_type(col, op);
             }
             break;
@@ -3988,9 +3993,9 @@ void emit_force(vcode_reg_t nets, vcode_reg_t nnets, vcode_reg_t values)
    vcode_add_arg(op, values);
 
    VCODE_ASSERT(vcode_reg_kind(nets) == VCODE_TYPE_SIGNAL,
-                "disconnect target is not signal");
+                "force target is not signal");
    VCODE_ASSERT(vcode_reg_kind(nnets) == VCODE_TYPE_OFFSET,
-                "disconnect net count is not offset type");
+                "force net count is not offset type");
 }
 
 void emit_release(vcode_reg_t nets, vcode_reg_t nnets)
@@ -4000,9 +4005,9 @@ void emit_release(vcode_reg_t nets, vcode_reg_t nnets)
    vcode_add_arg(op, nnets);
 
    VCODE_ASSERT(vcode_reg_kind(nets) == VCODE_TYPE_SIGNAL,
-                "disconnect target is not signal");
+                "release target is not signal");
    VCODE_ASSERT(vcode_reg_kind(nnets) == VCODE_TYPE_OFFSET,
-                "disconnect net count is not offset type");
+                "release net count is not offset type");
 }
 
 void emit_disconnect(vcode_reg_t nets, vcode_reg_t nnets, vcode_reg_t reject,
@@ -5441,6 +5446,24 @@ vcode_reg_t emit_link_package(ident_t name)
    op->ident = name;
 
    VCODE_ASSERT(name != active_unit->name, "cannot link the current unit");
+
+   return (op->result = vcode_add_reg(vtype_context(name)));
+}
+
+vcode_reg_t emit_link_instance(ident_t name, vcode_reg_t locus)
+{
+   VCODE_FOR_EACH_MATCHING_OP(other, VCODE_OP_LINK_INSTANCE) {
+      if (other->ident == name)
+         return other->result;
+   }
+
+   op_t *op = vcode_add_op(VCODE_OP_LINK_INSTANCE);
+   vcode_add_arg(op, locus);
+   op->ident = name;
+
+   VCODE_ASSERT(name != active_unit->name, "cannot link the current unit");
+   VCODE_ASSERT(vcode_reg_kind(locus) == VCODE_TYPE_DEBUG_LOCUS,
+                "locus argument to link instnance must be a debug locus");
 
    return (op->result = vcode_add_reg(vtype_context(name)));
 }