+++ NEWS.md
@@ -13,6 +13,7 @@
 - Arithmetic operations that overflow the underlying machine type now
   produce an error (#101).
 - Added support for VHDL-2008 force/release assignments.
+- Basic support for external names in VHDL-2008.
 
 ## Version 1.6.2 - 2022-04-03
 - Fix `make -j` with GNU make (#440).